+++ dejafu/Test/DejaFu/Conc.hs
@@ -7,7 +7,7 @@
 
 -- |
 -- Module      : Test.DejaFu.Conc
--- Copyright   : (c) 2016--2017 Michael Walker
+-- Copyright   : (c) 2016--2018 Michael Walker
 -- License     : MIT
 -- Maintainer  : Michael Walker <mike@barrucadu.co.uk>
 -- Stability   : experimental
@@ -359,20 +359,6 @@ dontCheck lb ma = toConc (ADontCheck lb (unC ma))
 --     pure r
 --   liftIO (readIORef r)
 -- @
-
--- | A snapshot of the concurrency state immediately after 'dontCheck'
--- finishes.
---
--- @since 1.1.0.0
-data DCSnapshot r n a = DCSnapshot
-  { dcsContext :: Context n r ()
-  -- ^ The execution context.  The scheduler state is ignored when
-  -- restoring.
-  , dcsRestore :: Threads n r -> n ()
-  -- ^ Action to restore CRef, MVar, and TVar values.
-  , dcsRef :: r (Maybe (Either Failure a))
-  -- ^ Reference where the result will be written.
-  }
 
 -- | Like 'runConcurrent', but terminates immediately after running
 -- the 'dontCheck' action with a 'DCSnapshot' which can be used in